Mark the correct ifaceImpl, add tests
In MarkMethod, we mark the interfaceImpl that has the same type
definition. If there are multiple implementations of a generic
interface, each with different type parameters, we would only mark the
first such interface. It just so happened that we never hit any issues
because elsewhere we would keep the interface implementations since the
interface definition was marked and the type was 'relevant to variant
casting'.

Adds tests for instance interface methods with .overrides in the
public implementations. It doesn't seems possible to create a scenario
where an instance method could be marked without the declaring type
being marked 'relevant to variant casting' and all the interfaces being
marked. Also, all methods pointed to by a .override is marked by
ProcessMethod, so there isn't a scenario that would have dangling
references in the .override.