Simplify the updating of high field
It also reduces the TP cost. The original thinking was branches can
be expensive and hence condition-less code was added, but we end up
with more instructions with condition-less. Additionally, branch predicators
should be better in making sure that mostly we would take the gpr/float code
path and rarely take the predicate registers code path. Also, we do PGO builds
which should feed that information and make branch code cheaper in TP cost wise.