Work around issue #102868
The relaxation of type layout check in #102810 exposed a bug in the type layout algorithm that doesn't correctly handle doing type layout in this situation (it doesn't match what CoreCLR VM computes). This restores the old behavior for ReadyToRun that just throws on this. The compiler will then stop precompiling the method and fall back to runtime JIT.

Works around #102868 (the bug still exists, this is a valid situation that we should be able to precompile, we just won't).