Make tests harder
On RISC-V the return value is passed like the first argument, so native compilers optimize out the passing and the test did not catch if the struct was passed incorrectly because it will be checking whatever intact registers were set on the call-site.

Besides, it's easier to see how the native compiler is passing arguments by disassembling the binary.