Fix CET - remove writing to shadow stack
In my recent fix for failures with CET enabled, I have also added a call to
the _wrssq intrinsic to push an address to shadow stack. It turns out
that instruction is privileged and cannot be used by user code.
Moreover, I have realized that it is not needed at all there, so I am
removing it.