[android] Fix build withe newer NDK
Newer versions of Android NDK (tested with 26.3) no longer include
support for 32-bit host machines and have moved libraries around for
that reason.  One of the libraries that changed its location is
`libclang.so`, required to build MonoVM for Android.

This PR adds support to detect the library at its new location.