Microsoft.IO.Redist part (and a build fix)
CHILD: .NET CORE | ElevationOfPrivilege: Microsoft.IO.Redist part (and a build fix)

It turned out that https://dev.azure.com/dnceng/internal/_git/dotnet-runtime/pullrequest/39614 was incomplete and caused a build break. The code was compiling fine for the default config (.NET 6), but not Full Framework, which includes Microsoft.IO.Redist which compiles the affected files.

I could not use exact same fix for Framework as for 6.0 as the code was using umanaged function pointers and causing error CS8889, so I've applied the same fix I did for Full Framework.