datas bug fixes (#103558)
we use dd_current_size of gen2 to compute the SOH stable size but during an HC change we update it incorrectly. because dd_current_size doesn't change inbetween GCs that condemn that generation, I just update it when we actually do a gen2 GC. note that this means the fix is incomplete because we do use it in other places. I'll fix those in a separate PR.

when we compute the gen0 DATAS budget, we need to use the 2 budgets as limits instead of not pick it at all if it doesn't land between those 2 budgets which causes us to just return the min of the 2. we can be in a situation where new_budget_per_heap is just slightly above budget_old_gen_per_heap but we end up picking a very small budget which is bcs_per_heap.