JIT: Avoid invalidating DFS tree unconditionally after RBO
Instead of invalidating the DFS tree unconditionally after RBO, we can
invalidate it only as part of RBO or assertion prop when they actually
make any flowgraph changes. The benefit is that when no flowgraph
changes are made, IV opts get access to the DFS tree, loops and
dominator tree without having to recompute them from scratch. This helps
on TP.

Making this work takes a bit of finesse since assertion prop has no
knowledge of when it actually made FG changes; instead we invalidate the
DFS tree directly from the morph function that can change the FG, with
the exception of when it is called from RBO, since RBO relies on the
dominator tree even after changing the flowgraph.

Also rename `fgGetDomSpeculatively` -> `optGetDomSpeculatively` and move
it into RBO, to make it more clear that it is only within RBO that
relying on the outdated dominators may be ok.