JIT: Use a hash table for HW intrinsic name -> ID lookups
This uses a lazily allocated global-memory hash table to perform these
lookups. I noticed that on benchmarks.run_pgo, this function was
disproportionately hot compared to how infrequently it is called; even
being called for only every 20th compilation on average, it accounted
for around 0.7% of JIT time. When jitting a method using a single HW
intrinsic the function accounted for 5.55% of JIT time. With this
change, that falls to 0.08%.