JIT: Specialize `fgFindJumpTargets` (#103761)
This specializes `fgFindJumpTargets` based on whether we are making
inline observations or not. This function has a lot of conditional logic
based on this property, but it is always going to be false in MinOpts,
so this allows the native compiler to optimize the function much better
(and avoids all the conditionals).

Using the VS profiler shows that the relative CPU time spent in the
function goes from 0.98% to 0.47% for the MinOpts contexts in
benchmarks.run_pgo.windows.x64.