Fix detection of lambda and local function generated names (#103916)
And share code between ILLink and Native AOT.

Lambdas and local functions inside explicitly-implemented interface
methods include the interface type in the generated name. When the
interface type is generic, this breaks our heuristics (which were
looking for the first `>` to indicate the end of the user method name
in the generated method). This updates the heuristic to look for the
last `>` instead.

Fixes https://github.com/dotnet/runtime/issues/98368