Show the linker cmake is using
This adds some extra line to the messages that cmake prints out:

    -- The C compiler identification is Clang 18.1.6
    ...
    -- The linker identification is LLD 18.1.6 (compatible with GNU linkers)
    ...

This makes it easier to see which linker is being used by the runtime.
This is useful in debugging, specially in scenarios where we pick the
wrong linker and then produce a broken runtime (like
https://github.com/dotnet/runtime/issues/43349), but have no quick way
to tell what linker was actually used.