Generalize IControlFlowGraph successors (#103676)
IControlFlowGraph was modeled after the Roslyn CFG API, where a block
can have at most two successors - a fall-through successor, and a
conditional successor. Switch statements are represented as a chain of
checks for one case at a time.

This doesn't generalize well to CFGs in IL which can have more than
two branch targets for a switch instruction. This changes the
interface to expose an IEnumerable<ControlFlowBranch> of successors.

- Move IsConditional on block to ConditionKind on branch

- Introduce GetSuccessors to replace GetConditionalSuccessor and
  GetFallThroughSuccessor