Enable sending activation to dispatch queue threads (#102887)
* Enable sending activation to dispatch queue threads

Until recently, it was not possible to send activation signal to Apple
dispatch queue threads using pthread_kill. So in case a .NET code was running
on such a thread, it could end up blocking runtime suspension in some cases.
Recently, Apple has implemented a new API to enable sending specific signals
to the dispatch queues, the dispatch_allow_send_signals.

This change adds a call to that API to the PAL initialization code.

* Workaround until CI can switch to newer xcode

* Add unmasking activation signal on foreign threads

The added coreclr test has revealed that while the signal could be sent now,
the inject_activation_handler was not being invoked on the dispatch queue
thread. The problem was that all signals are masked by default there.

* Add test to verify the fix

The test verifies that .NET code running on a dispatch queue thread in
a tight loop can get the activation injected.

* Fix build break

* The test should be Apple only

* Build the native part of the test on Apple only

* Fix reverse polarity assert in nativeaot

* Change usage of libdispatch to libSystem

* Make sure the testing is performed only on OS that supports the new API

* Add RequiresProcessIsolation for CLRTestTargetUnsupported