Disable tests with WaitForPendingFinalizers in gcstress testing
Some if not all tests from the Methodical suite that use GC.WaitForPendingFinalizers
timeout due to the fact that a finalizer of Gen2GcCallback calls a callback that
ends up creating another finalizable object that is ready to be finalized. Also
a GC memory is allocated in that callback, which triggers the GC stress collection.
So the finalization thread keeps finalizing objects on the queue forever and the
GC.WaitForPendingFinalizers never exits.

This change disables all tests in the Methodical suite that wait for pending
finalizers when running with GC stress enabled.