Fix access to hoisted variables from reductions
LambdaCompiler compiles several types of expressions by reducing them.
Reductions may contain BlockExpressions not processed by VariableBinder,
and in that case a scope is created for them on the fly during
compilation. In particular, SwitchExpression and TypeBinaryExpression
compilation can create a BlockExpression with a scope.

If such a BlockExpression is an immediate child of a LambdaExpression,
it lost access to lambda parameters that were hoisted due to the
propagation of the NeedsClosure flag for its scope from the lambda.

Always set the flag for such scopes to true, so that its scope always
gets access to hoisted variables from the parent scope.

Fix #56262