Revert ILLink's usage of dependency analysis framework (#104267)
See
https://github.com/dotnet/runtime/issues/103987#issuecomment-2197798838
for context.

Custom steps rely on getting a chance to see a type before we
build the type info (and in particular the interface method
mapping), but we make no such
guarantee. https://github.com/dotnet/runtime/issues/104266 tracks
this problem.

Our use of the dependency analysis framework exacerbated this
because `MarkType` was split into two pieces, with the part that
calls into the custom step being delayed through a dependency
analysis framework node, making it more likely to be run too late
to influence the type info.

This reverts ILLink's usage of the dependency analysis framework
to bring us back to a state where the ordering still isn't
guaranteed, but works for the testcase that got regressed. We
should bring this back as soon as possible with a proper fix that
actually guarantees the ordering required by custom steps. This
doesn't look completely straightforward, but should be possible
to do with the dependency analysis framework.

Fixes https://github.com/dotnet/runtime/issues/103987 (but we
need a better long-term fix for
https://github.com/dotnet/runtime/issues/104266).