Make CertLoader reject duplicate attributes
The new Pkcs12 loader will now reject duplicate attributes,
either as multiple attribute sets, or a set with multiple values.

The LoaderLimits type gains an option to disable this filter, but
as of now it is not being made into public API (though the tests
show how to get at it anyways, by the cloning behavior on
DangerousNoLimits).

This change also fixes the over-filtering of the MachineKey attribute,
and adds tests for how that behaves under DefaultKeySet.