Support using the system rapidjson to build (#101443)
We have a bundled copy of rapidjson in runtime, but some Linux
distributions also ship with another version of that library. Support
using that too, optionally.

This does not change the default behaviour - the bundled copy of
rapidjson is used. A cmake argument of
CLR_CMAKE_USE_SYSTEM_RAPIDJSON=true needs to be set to use the system
version of rapidjson.