fix the size passed to grow_heap_segment (#103998)
in loh_allocate_in_condemned we call grow_heap_segment with 2x loh_padding_obj_size while in loh_size_fit_p it specifically says one padding if it's at the end of the segment. while going from the amount of 1 padding to 2 isn't a big deal from grow_heap_segments POV it introduces an inconsistency between loh_size_fit_p and grow_heap_segment because the former will say fit but the latter will give it a size that does not fit. this triggers the assert -

assert (high_address <= heap_segment_reserved (seg));

because high_address is 0x20 (pad size) higher than reserved. in retail builds this would just return FALSE if we are right at the end of the segment which is also a problem.